Use lux count from new sensor, calibrate pressure,voc,co2

+++ suripu-core/src/main/java/com/hello/suripu/core/models/AllSensorSampleMap.java
@@ -53,7 +53,7 @@ public class AllSensorSampleMap {
                           final float co2,
                           final float ir,
                           final float clear,
-                          final int luxCount,
+                          final float luxCount,
                           final int uvCount) {
 
         put(Sensor.PRESSURE, dateTime, offsetMillis, pressure);
